# swagger-spec.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authorizaion API docs",
        "description": "The users API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "email",
                    "firstName",
                    "lastName",
                    "password",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "CreateUserDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "email of user"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "password_confirmation": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "email",
                    "firstName",
                    "lastName",
                    "password",
                    "password_confirmation",
                    "isActive"
                ]
            },
            "UpdateUserDto": {
                "type": "object",
                "properties": {}
            }
        }
    },
    "paths": {
        "/v1/users": {
            "get": {
                "operationId": "UserController_index",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "list user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ]
            },
            "post": {
                "operationId": "UserController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
...
        "/v1/users/{id}": {
            "get": {
                "operationId": "UserController_show",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ]
            },
            "put": {
                "operationId": "UserController_update",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ]
            },
            "delete": {
                "operationId": "UserController_destroy",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
...
    }
}